Refactor remove WebViewEventReportFragment.
