+++ app/src/main/java/org/hisp/dhis/android/dashboard/ui/activities/DashboardElementDetailActivity.java
@@ -47,7 +47,6 @@ import org.hisp.dhis.android.dashboard.api.models.Interpretation;
 import org.hisp.dhis.android.dashboard.api.models.InterpretationElement;
 import org.hisp.dhis.android.dashboard.api.models.InterpretationElement$Table;
 import org.hisp.dhis.android.dashboard.ui.fragments.ImageViewFragment;
-import org.hisp.dhis.android.dashboard.ui.fragments.WebViewEventReportFragment;
 import org.hisp.dhis.android.dashboard.ui.fragments.WebViewFragment;
 
 import butterknife.Bind;
@@ -144,14 +143,11 @@ public class DashboardElementDetailActivity extends BaseActivity {
                 attachFragment(ImageViewFragment.newInstance(request));
                 break;
             }
-            case DashboardItemContent.TYPE_REPORT_TABLE: {
-                String elementId = element.getUId();
-                attachFragment(WebViewFragment.newInstance(elementId));
-                break;
-            }
+            case DashboardItemContent.TYPE_REPORT_TABLE:
             case DashboardItemContent.TYPE_EVENT_REPORT: {
                 String elementId = element.getUId();
-                attachFragment(WebViewEventReportFragment.newInstance(elementId));
+                attachFragment(WebViewFragment.newInstance(elementId,
+                        element.getDashboardItem().getType()));
                 break;
             }
         }
@@ -176,7 +172,7 @@ public class DashboardElementDetailActivity extends BaseActivity {
             }
             case Interpretation.TYPE_REPORT_TABLE: {
                 String elementId = element.getUId();
-                attachFragment(WebViewFragment.newInstance(elementId));
+                attachFragment(WebViewFragment.newInstance(elementId, element.getType()));
                 break;
             }
             case Interpretation.TYPE_DATA_SET_REPORT: {
