+++ app/src/main/java/org/hisp/dhis/android/dashboard/ui/fragments/WebViewFragment.java
@@ -28,6 +28,10 @@
 
 package org.hisp.dhis.android.dashboard.ui.fragments;
 
+import static android.text.TextUtils.isEmpty;
+
+import static org.hisp.dhis.android.dashboard.api.models.DashboardItemContent.TYPE_REPORT_TABLE;
+
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -39,6 +43,8 @@ import org.hisp.dhis.android.dashboard.R;
 import org.hisp.dhis.android.dashboard.api.controllers.DhisController;
 import org.hisp.dhis.android.dashboard.api.job.Job;
 import org.hisp.dhis.android.dashboard.api.job.JobExecutor;
+import org.hisp.dhis.android.dashboard.api.models.DataElementDimension;
+import org.hisp.dhis.android.dashboard.api.models.EventReport;
 import org.hisp.dhis.android.dashboard.api.models.meta.ResponseHolder;
 import org.hisp.dhis.android.dashboard.api.network.APIException;
 import org.hisp.dhis.android.dashboard.api.network.DhisApi;
@@ -48,15 +54,16 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.lang.ref.WeakReference;
+import java.util.ArrayList;
+import java.util.List;
 
 import butterknife.Bind;
 import butterknife.ButterKnife;
 import retrofit.mime.TypedInput;
 
-import static android.text.TextUtils.isEmpty;
-
 public class WebViewFragment extends BaseFragment {
     private static final String DASHBOARD_ELEMENT_ID = "arg:dashboardElementId";
+    private static final String DASHBOARD_TYPE = "dashboardType";
 
     @Bind(R.id.web_view_content)
     WebView mWebView;
@@ -64,9 +71,10 @@ public class WebViewFragment extends BaseFragment {
     @Bind(R.id.container_layout_progress_bar)
     View mProgressBarContainer;
 
-    public static WebViewFragment newInstance(String id) {
+    public static WebViewFragment newInstance(String id, String dashboardType) {
         Bundle args = new Bundle();
         args.putString(DASHBOARD_ELEMENT_ID, id);
+        args.putString(DASHBOARD_TYPE, dashboardType);
 
         WebViewFragment fragment = new WebViewFragment();
         fragment.setArguments(args);
@@ -84,9 +92,16 @@ public class WebViewFragment extends BaseFragment {
 
         mWebView.getSettings().setBuiltInZoomControls(true);
         if (getArguments() != null && !isEmpty(getArguments()
-                .getString(DASHBOARD_ELEMENT_ID))) {
-            JobExecutor.enqueueJob(new GetReportTableJob(this, getArguments()
-                    .getString(DASHBOARD_ELEMENT_ID)));
+                .getString(DASHBOARD_ELEMENT_ID)) && !isEmpty(getArguments()
+                .getString(DASHBOARD_TYPE))) {
+            if (getArguments()
+                    .getString(DASHBOARD_TYPE).equals(TYPE_REPORT_TABLE)) {
+                JobExecutor.enqueueJob(new GetReportTableJob(this, getArguments()
+                        .getString(DASHBOARD_ELEMENT_ID)));
+            } else {
+                JobExecutor.enqueueJob(new GetEventReportTableJob(this, getArguments()
+                        .getString(DASHBOARD_ELEMENT_ID)));
+            }
         }
     }
 
@@ -157,6 +172,48 @@ public class WebViewFragment extends BaseFragment {
             if (mFragmentRef.get() != null) {
                 mFragmentRef.get().onDataDownloaded(result);
             }
+        }
+    }
+
+    static class GetEventReportTableJob extends GetReportTableJob {
+        public GetEventReportTableJob(WebViewFragment fragment, String dashboardElementId) {
+            super(fragment, dashboardElementId);
+        }
+
+        @Override
+        public ResponseHolder<String> inBackground() {
+            ResponseHolder<String> responseHolder = new ResponseHolder<>();
+            EventReport eventReport;
+
+            try {
+                DhisApi dhisApi = RepoManager.createService(
+                        DhisController.getInstance().getServerUrl(),
+                        DhisController.getInstance().getUserCredentials());
+                eventReport = dhisApi.getEventReport(mDashboardElementId);
+                responseHolder.setItem(readInputStream(
+                        dhisApi.getEventReportTableData(eventReport.getProgram().getuId(),
+                                eventReport.getProgramStage().getuId(),
+                                getDimensions(eventReport)).getBody()));
+            } catch (APIException exception) {
+                responseHolder.setApiException(exception);
+            }
+
+            return responseHolder;
+        }
+
+        private List<String> getDimensions(EventReport eventReport) {
+            List<String> dimensions = new ArrayList<>();
+            dimensions.add("pe:" + eventReport.getRelativePeriods().getRelativePeriodString());
+            dimensions.add("ou:" + eventReport.getOrganisationUnits().get(0).getuId());
+            for (DataElementDimension dimension : eventReport.getDataElementDimensions()) {
+                String dimensionUID = "";
+                dimensionUID += dimension.getDataElement().getuId();
+                if (dimension.getFilter() != null && !dimension.getFilter().isEmpty()) {
+                    dimensionUID += ":" + dimension.getFilter();
+                }
+                dimensions.add(dimensionUID);
+            }
+            return dimensions;
         }
     }
 }
